[#7] [UI] Move LoginForm to a new file, rename package, and fix TextFormField colors
