+++ lib/ui/login/login_screen.dart
@@ -1,12 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-
 import '../../gen/assets.gen.dart';
-import '../../theme/app_colors.dart';
 import '../widget/dimmed_image_background.dart';
-import '../widget/primary_text_form_field_decoration.dart';
+import '../widget/login_form.dart';
 
 const _logoRevealDuration = Duration(milliseconds: 500);
 const _logoDuration = Duration(milliseconds: 750);
@@ -118,111 +115,5 @@ class _LoginScreenState extends ConsumerState<LoginScreen>
   void dispose() {
     _logoRevealAnimationController.dispose();
     super.dispose();
-  }
-}
-
-class LoginForm extends ConsumerStatefulWidget {
-  const LoginForm({super.key});
-
-  @override
-  LoginFormState createState() => LoginFormState();
-}
-
-class LoginFormState extends ConsumerState<LoginForm> {
-  final _formKey = GlobalKey<FormState>();
-
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-      key: _formKey,
-      autovalidateMode: AutovalidateMode.onUserInteraction,
-      child: _buildLoginForm(),
-    );
-  }
-
-  Widget _buildLoginForm() {
-    const forgotButtonWidth = 80.0;
-
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        TextFormField(
-          decoration: PrimaryTextFormFieldDecoration(
-            context: context,
-            hint: AppLocalizations.of(context)!.email,
-          ),
-          keyboardType: TextInputType.emailAddress,
-          style: Theme.of(context).textTheme.bodyLarge,
-          textInputAction: TextInputAction.next,
-        ),
-        const SizedBox(height: 20),
-        Stack(
-          children: [
-            TextFormField(
-              decoration: PrimaryTextFormFieldDecoration(
-                context: context,
-                hint: AppLocalizations.of(context)!.password,
-              ).copyWith(
-                  contentPadding: const EdgeInsets.only(
-                      left: 12,
-                      right: forgotButtonWidth,
-                      top: 15, // O:
-                      bottom: 15 // use dimen constants
-                      )),
-              obscureText: true,
-              obscuringCharacter: "●",
-              style: Theme.of(context).textTheme.bodyLarge,
-              textInputAction: TextInputAction.done,
-              onFieldSubmitted: (_) => {
-                // TODO: log in
-              },
-            ),
-            Align(
-              alignment: Alignment.centerRight,
-              child: SizedBox(
-                height: 56,
-                child: TextButton(
-                  child: Text(
-                    AppLocalizations.of(context)!.forgotPassword,
-                    style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                          color: AppColors.whiteAlpha50,
-                        ),
-                  ),
-                  onPressed: () {
-                    // TODO: navigate to reset password
-                  },
-                ),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 20),
-        SizedBox(
-          width: double.infinity,
-          child: TextButton(
-            style: ButtonStyle(
-              backgroundColor: MaterialStateProperty.all(Colors.white),
-              foregroundColor: MaterialStateProperty.all(Colors.black),
-              overlayColor: MaterialStateProperty.all(Colors.black12),
-              padding: MaterialStateProperty.all(
-                const EdgeInsets.symmetric(
-                  vertical: 12,
-                ),
-              ),
-              shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-                RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-              ),
-              textStyle: MaterialStateProperty.all(
-                Theme.of(context).textTheme.labelLarge,
-              ),
-            ),
-            child: Text(AppLocalizations.of(context)!.login),
-            onPressed: () => {
-              // TODO: log in
-            },
-          ),
-        ),
-      ],
-    );
   }
 }
